backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
cryptography>=41.0.0
pandas>=2.0.0
pandas-ta
yfinance>=0.2.28
anthropic>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic-settings>=2.0.3
apscheduler>=3.10.1
pytz>=2023.3
passlib[bcrypt]
PyJWT
openpyxl>=3.1.0
